refactor: change convert moduule to utils
