+++ src/ape_accounts/accounts.py
@@ -7,9 +7,9 @@ from eth_account import Account as EthAccount  # type: ignore
 from eth_utils import to_bytes
 
 from ape.api import AccountAPI, AccountContainerAPI, TransactionAPI
-from ape.convert import to_address
 from ape.exceptions import AccountsError
 from ape.types import AddressType, MessageSignature, SignableMessage, TransactionSignature
+from ape.utils import to_address
 
 
 class InvalidPasswordError(AccountsError):
